+++ LR1/app/src/main/java/com/example/unitconverter/model/ConversionResult.kt
@@ -1,11 +1,13 @@
 package com.example.unitconverter.model
 
+import java.math.BigDecimal
+
 /**
  * Represents the result of a unit conversion
  */
 data class ConversionResult(
-    val inputValue: Double,
-    val outputValue: Double,
+    val inputValue: BigDecimal,
+    val outputValue: BigDecimal,
     val fromUnit: UnitItem,
     val toUnit: UnitItem,
     val formattedResult: String = ""
